+++ scripts/eject.mjs
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import { fileURLToPath } from 'url'
 import { dirname } from 'path'
-import { deleteFile, loadFileData } from './utils'
+import { deleteFile, loadFileData } from './utils.mjs'
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
